Færdig med recursion BinarySearch  ex 4

+++ i3doa-recursion/ClassroomExercises/ClassroomExercises.h
@@ -2,6 +2,6 @@
 
 unsigned int factorial(unsigned int n);
 void printDownTo1(unsigned int n);
-double power(int x, int exp);
+double power(double x, double exp);
 unsigned int fib(unsigned int n);
 void hanoi(int nDisks, int from, int to, int via);
